+++ packages/colony-js-adapter/interface/Adapter.js
@@ -3,7 +3,6 @@
 import ContractLoader from '@colony/colony-js-contract-loader';
 import type { Query } from '@colony/colony-js-contract-loader';
 import type { Contract } from './Contract';
-import type { EventHandlers } from './EventHandlers';
 import type { Provider } from './Provider';
 import type { Signature } from './Signature';
 import type { Transaction } from './Transaction';
@@ -20,11 +19,6 @@ export interface Adapter {
     query: Query,
     contractParams: Array<any>,
   ): Promise<Transaction>;
-  getEventData({
-    events: EventHandlers,
-    timeoutMs: number,
-    transactionHash: string,
-  }): Promise<any>;
   waitForTransaction(
     transactionHash: string,
     timeoutMs?: number,