+++ packages/colony-js-adapter/interface/Contract.js
@@ -2,7 +2,6 @@
 
 import type BigNumber from 'bn.js';
 
-import type { Event } from './Event';
 import type { EventCallback } from './EventHandlers';
 import type { Transaction } from './Transaction';
 import type { TransactionOptions } from './TransactionOptions';
@@ -16,6 +15,14 @@ export interface Contract {
     [string]: (Function) => void,
   };
   connect: Wallet => Contract;
+  interface: {
+    events: {
+      [eventName: string]: {
+        parse(topics: Array<string>, data: Object): Object,
+        topics: Array<string>,
+      },
+    },
+  };
   addListener(
     eventName: string,
     callback: EventCallback,