+++ packages/colony-js-client/src/__tests__/ColonyNetworkClient.test.js
@@ -2,7 +2,6 @@
 
 import { utf8ToHex } from 'web3-utils';
 import createSandbox from 'jest-sandbox';
-import BigNumber from 'bn.js';
 import ColonyNetworkClient from '../ColonyNetworkClient';
 import ColonyClient from '../ColonyClient';
 import TokenClient from '../TokenClient';
@@ -84,21 +83,9 @@ describe('ColonyNetworkClient', () => {
       networkClient.constructor.Sender,
     );
     expect(networkClient.createColony).toHaveProperty('_defaultGasLimit');
-    expect(networkClient.createColony.eventHandlers).toEqual({
-      ColonyAdded: {
-        contract: {},
-        handler: expect.any(Function),
-      },
-    });
     expect(networkClient.createColony.input).toEqual([
       ['tokenAddress', 'address'],
     ]);
-    expect(
-      networkClient.createColony.eventHandlers.ColonyAdded.handler({
-        colonyId: new BigNumber(100),
-        colonyAddress: 'colony address',
-      }),
-    ).toEqual({ colonyId: 100, colonyAddress: 'colony address' });
 
     const response = {
       eventData: {