+++ packages/colony-js-contract-client/src/classes/ContractMethodSender.js
@@ -4,7 +4,6 @@ import BigNumber from 'bn.js';
 import { raceAgainstTimeout } from '@colony/colony-js-utils';
 
 import type {
-  EventHandlers,
   Transaction,
   TransactionOptions,
 } from '@colony/colony-js-adapter';
@@ -22,20 +21,16 @@ export default class ContractMethodSender<
   OutputValues: { [outputValueName: string]: any },
   IContractClient: ContractClient,
 > extends ContractMethod<InputValues, OutputValues, IContractClient> {
-  eventHandlers: EventHandlers;
   _defaultGasLimit: ?number;
 
   constructor({
     defaultGasLimit,
-    eventHandlers,
     ...rest
   }: ContractMethodArgs<IContractClient> & {
-    eventHandlers?: EventHandlers,
     defaultGasLimit?: number,
   }) {
     super(rest);
     if (defaultGasLimit) this._defaultGasLimit = defaultGasLimit;
-    if (eventHandlers) this.eventHandlers = eventHandlers;
   }
 
   /**
@@ -65,15 +60,11 @@ export default class ContractMethodSender<
     transaction: Transaction,
     timeoutMs: number,
   ): Promise<ContractResponse<OutputValues>> {
-    const eventData = await this.client.getEventData({
-      events: this.eventHandlers,
-      timeoutMs,
-      transactionHash: transaction.hash,
-    });
     const receipt = await raceAgainstTimeout(
       this.client.adapter.getTransactionReceipt(transaction.hash),
       timeoutMs,
     );
+    const eventData = await this.client.getReceiptEventData(receipt);
 
     return {
       successful: receipt && receipt.status === 1,
@@ -89,15 +80,12 @@ export default class ContractMethodSender<
     transaction: Transaction,
     timeoutMs: number,
   ): ContractResponse<OutputValues> {
-    const eventDataPromise = this.client.getEventData({
-      events: this.eventHandlers,
-      timeoutMs,
-      transactionHash: transaction.hash,
-    });
     const receiptPromise = raceAgainstTimeout(
       this.client.adapter.getTransactionReceipt(transaction.hash),
       timeoutMs,
     );
+
+    // Wait for the receipt before determining whether it was successful
     const successfulPromise = new Promise(async (resolve, reject) => {
       try {
         const receipt = await receiptPromise;
@@ -106,6 +94,21 @@ export default class ContractMethodSender<
         reject(error.toString());
       }
     });
+
+    // Wait for the receipt before attempting to decode event logs
+    const eventDataPromise = new Promise(async (resolve, reject) => {
+      try {
+        const receipt = await receiptPromise;
+        try {
+          resolve(await this.client.getReceiptEventData(receipt));
+        } catch (decodeError) {
+          reject(decodeError.toString());
+        }
+      } catch (receiptError) {
+        reject(receiptError.toString());
+      }
+    });
+
     return {
       successfulPromise,
       meta: {